Crate reth_rpc_types

Reth RPC type definitions.

Provides all relevant types for the various RPC endpoints, grouped by namespace.

Re-exports

pub use eth::engine;

Modules

beacon

Types for the Ethereum 2.0 RPC protocol (beacon chain).

error

Commonly used errors for the eth_ namespace.

kzg

KZG type bindings

optimism

Misc Optimism-specific types

other

Support for capturing other fields

pubsub

Ethereum types for pub-sub

raw_log

Ethereum log object.

relay

Relay API bindings: https://flashbots.github.io/relay-specs/

request

Alloy basic Transaction Request type.

serde_helpers

Serde helpers for primitive types.

state

bindings for state overrides in eth_call

trace

RPC types for trace endpoints and inspectors.

transaction

RPC types for transactions

txpool

Types for the txpool namespace: https://geth.ethereum.org/docs/interacting-with-geth/rpc/ns-txpool

withdrawal

Withdrawal type and serde helpers.

Structs

AccessList

AccessList as defined in EIP-2930

AccessListItem

A list of addresses and storage keys that the transaction plans to access. Accesses outside the list are possible, but become more expensive.

AccessListWithGasUsed

Access list with gas used appended.

AccountInfo

Account information.

BlobTransactionSidecar

This represents a set of blobs, and its corresponding commitments and proofs.

Block

Block representation

BlockDetails

Custom struct for otterscan getBlockDetails RPC response

BlockNumHash

Block number and hash.

BlockOverrides

BlockOverrides is a set of header fields to override.

BlockTransactionHashes

An iterator over the transaction hashes of a block.

BlockTransactionHashesMut

An Iterator over the transaction hashes of a block.

BloomFilter

Helper type to represent a bloom filter used for matching logs.

Bundle

Bundle of transactions

CallInput

Helper type that supports both data and input fields that map to transaction input data.

CallInputError

Error thrown when both data and input fields are set and not equal.

CallRequest

Call request for eth_call and adjacent methods.

CancelBundleRequest

Request for eth_cancelBundle

CancelPrivateTransactionRequest

Request for eth_cancelPrivateTransaction

ChainStatus

Chain status.

ConsideredByBuildersAt

Represents information about when a bundle was considered by a builder.

ContractCreator

Custom struct for otterscan getContractCreator RPC responses

EIP1186AccountProofResponse

Response for EIP-1186 account proof eth_getProof

EIP1186StorageProof

Data structure with proof for one single storage-entry

EthBundleHash

Response from the matchmaker after sending a bundle.

EthCallBundle

Bundle of transactions for eth_callBundle

EthCallBundleResponse

Response for eth_callBundle

EthCallBundleTransactionResult

Result of a single transaction in a bundle for eth_callBundle

EthCallResponse

CallResponse for eth_callMany

EthProtocolInfo

Information about the Ethereum Wire Protocol (ETH)

EthSendBundle

Bundle of transactions for eth_sendBundle

ExecutionPayloadV1

This structure maps on the ExecutionPayload structure of the beacon chain spec.

ExecutionPayloadV2

This structure maps on the ExecutionPayloadV2 structure of the beacon chain spec.

ExecutionPayloadV3

This structure maps on the ExecutionPayloadV3 structure of the beacon chain spec.

ExtAccountInfo

Extended account information (used by parity_allAccountInfo).

FeeHistory

Response type for eth_feeHistory

Filter

Filter for logs.

FilterSet

FilterSet is a set of values that will be used to filter logs

FilteredParams

Support for matching Filters

Header

Block header representation.

HexStringMissingPrefixError

Thrown when a 0x-prefixed hex string was expected

Inclusion

Data used by block builders to check if the bundle should be considered for inclusion.

Index

A hex encoded or decimal index that’s intended to be used as a rust index, hence it’s deserialized into a usize.

InternalIssuance

Internal issuance struct for BlockDetails struct

InternalOperation

Custom struct for otterscan getInternalOperations RPC response

JsonStorageKey

A storage key type that can be serialized to and from a hex string up to 32 bytes. Used for eth_getStorageAt and eth_getProof RPCs.

JsonU256

Wrapper around primitive U256 type that also supports deserializing numbers

Log

Ethereum Log emitted by a transaction

NetworkStatus

The status of the network being ran by the local node.

NodeInfo

Represents the admin_nodeInfo response, which can be queried for all the information known about the running node at the networking granularity.

NodeRecord

Represents a ENR in discovery.

OptimismTransactionReceiptFields

Additional fields for Optimism transaction receipts

OtsBlock

Custom Block struct that includes transaction count for Otterscan responses

OtsBlockTransactions

Custom struct for otterscan getBlockTransactions RPC response

OtsTransactionReceipt

Custom transaction receipt struct for otterscan OtsBlockTransactions struct

Parity

Type that represents the signature parity byte, meant for use in RPC.

ParseBlockHashOrNumberError

Error thrown when parsing a BlockHashOrNumber from a string.

PeerEthProtocolInfo

Peer Ethereum protocol information

PeerInfo

Peer connection information

PeerNetworkInfo

Peer network information

PeerProtocolsInfo

Peer protocols information

Peers

Peers info

PipProtocolInfo

Peer PIP protocol information

Ports

Ports exposed by the node for discovery and listening.

Privacy

Preferences on what data should be shared about the bundle and its transactions

PrivacyHint

Hints on what data should be shared about the bundle and its transactions

PrivateTransactionPreferences

Additional preferences for eth_sendPrivateTransaction

PrivateTransactionRequest

Request for eth_sendPrivateTransaction

Protocols

All supported protocols

RawLog

Ethereum Log

RecoveredAccount

account derived from a signature as well as information that tells if it is valid for the current chain

Refund

Specifies the minimum percent of a given bundle’s earnings to redistribute for it to be included in a builder’s block.

RefundConfig

Specifies what addresses should receive what percent of the overall refund for this bundle, if it is enveloped by another bundle (eg. a searcher backrun).

Rich

Value representation with additional info

RpcBlockHash

A block hash which may have a boolean requireCanonical field. If false, an RPC call should raise if a block matching the hash is not found. If true, an RPC call should additionaly raise if the block is not in the canonical chain. https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md#specification

RpcModules

Represents the rpc_modules response, which returns the list of all available modules on that transport and their version

SealedByBuildersAt

Represents information about when a bundle was sealed by a builder.

SendBundleRequest

A bundle of transactions to send to the matchmaker.

SendBundleResponse

Response from the matchmaker after sending a bundle.

Signature

Container type for all signature fields in RPC

SimBundleLogs

Logs returned by mev_simBundle.

SimBundleOverrides

Optional fields to override simulation state.

SimBundleResponse

Response from the matchmaker after sending a simulation request.

StateContext

State context for callMany

StatsSeen

Response for flashbots_getBundleStatsV2 represents stats for a single bundle

StatsSimulated

Response for flashbots_getBundleStatsV2 represents stats for a single bundle

SyncInfo

Syncing info

TraceEntry

Custom struct for otterscan traceTransaction RPC response

Transaction

Transaction object used in RPC

TransactionInfo

Additional fields in the context of a block that contains this transaction.

TransactionReceipt

Transaction receipt

TransactionRequest

Represents all transaction requests received from RPC

TransactionStats

Propagation statistics for pending transaction.

TransactionsWithReceipts

Custom struct for otterscan searchTransactionsAfterand searchTransactionsBefore RPC responses

TxGasAndReward

Internal struct to calculate reward percentiles

U64HexOrNumber

A u64 wrapper type that deserializes from hex or a u64 and serializes as hex.

UserStats

Response for flashbots_getUserStatsV2 represents stats for a searcher.

Validity

Requirements for the bundle to be included in the block.

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Work

The result of an eth_getWork request

Enums

BlockError

Error that can occur when converting other types to blocks

BlockHashOrNumber

Either a block hash or a block number

BlockId

A Block Identifier https://github.com/ethereum/EIPs/blob/master/EIPS/eip-1898.md

BlockNumberOrTag

A block Number (or tag - “latest”, “earliest”, “pending”)

BlockTransactions

Block Transactions depending on the boolean attribute of eth_getBlockBy*, or if used by eth_getUncle*

BlockTransactionsKind

Determines how the transactions field of Block should be filled.

BundleItem

A bundle tx, which can either be a transaction hash, or a full tx.

BundleStats

Response for flashbots_getBundleStatsV2 represents stats for a single bundle

ExecutionPayload

An execution payload, which can be either ExecutionPayloadV1, ExecutionPayloadV2, or ExecutionPayloadV3.

FilterBlockOption

Represents the target range of blocks for the filter

FilterChanges

Response of the eth_getFilterChanges RPC.

FilterId

Owned equivalent of a SubscriptionId

LogError

Error that can occur when converting other types to logs

NodeRecordParseError

Possible error types when parsing a NodeRecord

NumberOrHexU256

An enum that represents either a serde_json::Number integer, or a hex U256.

OperationType

Operation type enum for InternalOperation struct

ParseBlockNumberError

Error variants when parsing a BlockNumberOrTag

PayloadError

Error that can occur when handling payloads.

PeerCount

Number of peers connected to.

PendingTransactionFilterKind

Specifies the kind of information you wish to receive from the eth_newPendingTransactionFilter RPC endpoint.

ProtocolVersion

The version of the MEV-share API to use.

SyncStatus

Sync status

TransactionKind

Represents the to field of a transaction request

TypedTransactionRequest

Container type for various Ethereum transaction requests

ValueOrArray

Union type for representing a single value or a vector of values inside a filter

Functions

from_int_or_hex

Deserializes the input into a U256, accepting both 0x-prefixed hex and decimal strings with arbitrary precision, defined by serde_json’s Number.

from_int_or_hex_opt

Deserializes the input into an Option<U256>, using from_int_or_hex to deserialize the inner value.

logs_bloom

Calculate receipt logs bloom.

serialize_b256_hex_string_no_prefix

Serialize a B256 as a hex string without the “0x” prefix.

serialize_hex_string_no_prefix

Serialize a byte vec as a hex string without the “0x” prefix.

Type Aliases

ForkBlock

Block number and hash of the forked block.

PeerId

Alias for a peer identifier

RichBlock

A Block representation that allows to include additional fields

RichHeader

Header representation with additional info.

Topic

A single topic