Module reth_rpc_types::beacon::events

Beacon API events support. Support for the Beacon API events

See also ethereum-beacon-API eventstream

Modules

attestation

light_client_finality

light_client_optimistic

Structs

AttestationEvent

Event for the Attestation topic of the beacon API node event stream.

BlobSidecarEvent

Event for the BlobSidecar topic of the beacon API node event stream.

BlockEvent

Event for the Block topic of the beacon API node event stream.

BlsToExecutionChangeEvent

Event for the BlsToExecutionChange topic of the beacon API node event stream.

BlsToExecutionChangeMessage

ChainReorgEvent

Event for the ChainReorg topic of the beacon API node event stream.

Contribution

ContributionAndProofEvent

Event for the ContributionAndProof topic of the beacon API node event stream.

ContributionAndProofMessage

FinalizedCheckpointEvent

Event for the Deposit topic of the beacon API node event stream.

HeadEvent

Event for the Head topic of the beacon API node event stream.

LightClientFinalityUpdateEvent

Event for the LightClientFinalityUpdate topic of the beacon API node event stream.

LightClientOptimisticUpdateEvent

Event for the LightClientOptimisticUpdate topic of the beacon API node event stream.

PayloadAttributesData

Data of the event that contains the payload attributes

PayloadAttributesEvent

Event for the payload_attributes topic of the beacon API node event stream.

VoluntaryExitEvent

Event for the VoluntaryExit topic of the beacon API node event stream.

VoluntaryExitMessage

Enums

BeaconNodeEventTopic

Topic variant for the eventstream API