Module reth_rpc_types::beacon::payload

Payload support for the beacon API.

Internal helper module to deserialize/serialize the payload attributes for the beacon API, which uses snake case and quoted decimals.

This is necessary because we don’t want to allow a mixture of both formats, hence serde aliases are not an option.

See also https://github.com/ethereum/consensus-specs/blob/master/specs/deneb/beacon-chain.md#executionpayload

Modules

beacon_api_payload_attributes

A helper module for serializing and deserializing the payload attributes for the beacon API.

beacon_api_payload_attributes_optimism

A helper module for serializing and deserializing optimism payload attributes for the beacon API.

beacon_payload

beacon_payload_v1

A helper serde module to convert from/to the Beacon API which uses quoted decimals rather than big-endian hex.

beacon_payload_v2

A helper serde module to convert from/to the Beacon API which uses quoted decimals rather than big-endian hex.

beacon_payload_v3

A helper serde module to convert from/to the Beacon API which uses quoted decimals rather than big-endian hex.

Structs

ExecutionPayloadHeader

The header of the execution payload.

ExecutionPayloadHeaderData

ExecutionPayloadHeaderMessage

GetExecutionPayloadHeaderResponse

Response object of GET /eth/v1/builder/header/{slot}/{parent_hash}/{pubkey}