Module reth_rpc_types::engine

Engine API types: https://github.com/ethereum/execution-apis/blob/main/src/engine/authentication.md and https://eips.ethereum.org/EIPS/eip-3675, following the execution specs https://github.com/ethereum/execution-apis/tree/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine.

Modules

payload

Structs

BlobsBundleV1

This includes all bundled blob related data of an executed payload.

CancunPayloadFields

Fields introduced in engine_newPayloadV3 that are not present in the ExecutionPayload RPC object.

ExecutionPayloadBodyV1

This structure contains a body of an execution payload.

ExecutionPayloadEnvelopeV2

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V2.

ExecutionPayloadEnvelopeV3

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V3.

ExecutionPayloadInputV2

This is the input to engine_newPayloadV2, which may or may not have a withdrawals field.

ExecutionPayloadV1

This structure maps on the ExecutionPayload structure of the beacon chain spec.

ExecutionPayloadV2

This structure maps on the ExecutionPayloadV2 structure of the beacon chain spec.

ExecutionPayloadV3

This structure maps on the ExecutionPayloadV3 structure of the beacon chain spec.

ForkchoiceState

This structure encapsulates the fork choice state

ForkchoiceUpdated

Represents a successfully processed forkchoice state update.

MaybeCancunPayloadFields

A container type for CancunPayloadFields that may or may not be present.

OptimismPayloadAttributes

Optimism Payload Attributes

PayloadAttributes

This structure contains the attributes required to initiate a payload build process in the context of an engine_forkchoiceUpdated call.

PayloadId

And 8-byte identifier for an execution payload.

PayloadStatus

This structure contains the result of processing a payload or fork choice update.

TransitionConfiguration

This structure contains configurable settings of the transition process.

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Enums

ExecutionPayload

An execution payload, which can be either ExecutionPayloadV1, ExecutionPayloadV2, or ExecutionPayloadV3.

ExecutionPayloadFieldV2

This represents the executionPayload field in the return value of engine_getPayloadV2, specified as:

ForkchoiceUpdateError

A standalone forkchoice update errors for RPC.

PayloadError

Error that can occur when handling payloads.

PayloadStatusEnum

PayloadValidationError

Various errors that can occur when validating a payload or forkchoice update.

Constants

CAPABILITIES

The list of all supported Engine capabilities available over the engine endpoint.

INVALID_FORK_CHOICE_STATE_ERROR

invalid forkchoice state error code.

INVALID_FORK_CHOICE_STATE_ERROR_MSG

invalid forkchoice state error message.

INVALID_PAYLOAD_ATTRIBUTES_ERROR

invalid payload attributes error code.

INVALID_PAYLOAD_ATTRIBUTES_ERROR_MSG

invalid payload attributes error message.

Type Aliases

ExecutionPayloadBodiesV1

The execution payload body response that allows for null values.

ForkChoiceUpdateResult

Represents possible variants of a processed forkchoice update.