Module reth_rpc_types::engine::payload

Structs

BlobsBundleV1

This includes all bundled blob related data of an executed payload.

ExecutionPayloadBodyV1

This structure contains a body of an execution payload.

ExecutionPayloadEnvelopeV2

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V2.

ExecutionPayloadEnvelopeV3

This structure maps for the return value of engine_getPayload of the beacon chain spec, for V3.

ExecutionPayloadInputV2

This is the input to engine_newPayloadV2, which may or may not have a withdrawals field.

ExecutionPayloadV1

This structure maps on the ExecutionPayload structure of the beacon chain spec.

ExecutionPayloadV2

This structure maps on the ExecutionPayloadV2 structure of the beacon chain spec.

ExecutionPayloadV3

This structure maps on the ExecutionPayloadV3 structure of the beacon chain spec.

PayloadAttributes

This structure contains the attributes required to initiate a payload build process in the context of an engine_forkchoiceUpdated call.

PayloadId

And 8-byte identifier for an execution payload.

PayloadStatus

This structure contains the result of processing a payload or fork choice update.

Withdrawal

Withdrawal represents a validator withdrawal from the consensus layer.

Enums

ExecutionPayload

An execution payload, which can be either ExecutionPayloadV1, ExecutionPayloadV2, or ExecutionPayloadV3.

ExecutionPayloadFieldV2

This represents the executionPayload field in the return value of engine_getPayloadV2, specified as:

PayloadError

Error that can occur when handling payloads.

PayloadStatusEnum

PayloadValidationError

Various errors that can occur when validating a payload or forkchoice update.

Type Aliases

ExecutionPayloadBodiesV1

The execution payload body response that allows for null values.