Module reth_rpc_types::relay

Relay API bindings: https://flashbots.github.io/relay-specs/

Modules

error

Error types for the relay.

Structs

BidTrace

Represents public information about a block sent by a builder to the relay, or from the relay to the proposer. Depending on the context, value might represent the claimed value by a builder (not necessarily a value confirmed by the relay).

BuilderBlockValidationRequest

A Request to validate a SubmitBlockRequest https://github.com/flashbots/builder/blob/03ee71cf0a344397204f65ff6d3a917ee8e06724/eth/block-validation/api.go#L132-L136

BuilderBlockValidationRequestV2

A Request to validate a SubmitBlockRequest https://github.com/flashbots/builder/blob/03ee71cf0a344397204f65ff6d3a917ee8e06724/eth/block-validation/api.go#L204-L204

BuilderBlocksReceivedQuery

Query for the GET /relay/v1/data/bidtraces/builder_blocks_received endpoint. This endpoint provides BidTraces for builder block submissions that match the query and were verified successfully.

ProposerPayloadsDeliveredQuery

Query for the GET /relay/v1/data/bidtraces/proposer_payload_delivered

SignedBidSubmissionV1

Submission for the /relay/v1/builder/blocks endpoint (Bellatrix).

SignedBidSubmissionV2

Submission for the /relay/v1/builder/blocks endpoint (Capella).

SignedBidSubmissionV3

Submission for the /relay/v1/builder/blocks endpoint (Deneb).

SignedBidTrace

SignedBidTrace is a BidTrace with a signature

SubmitBlockRequest

SubmitBlockRequest is the request from the builder to submit a block.

Validator

Represents an entry of the /relay/v1/builder/validators endpoint

ValidatorRegistration

Details of a validator registration.

ValidatorRegistrationMessage

Represents the message of a validator registration.

Enums

OrderBy

OrderBy : Sort results in either ascending or descending values. * -value - descending value (highest value first) * value - ascending value (lowest value first) Sort results in either ascending or descending values. * -value - descending value (highest value first) * value