Module reth_rpc_types::serde_helpers::storage

Storage related helpers.

Structs

JsonStorageKey

A storage key type that can be serialized to and from a hex string up to 32 bytes. Used for eth_getStorageAt and eth_getProof RPCs.

Functions

deserialize_storage_map

Deserializes the input into an Option<HashMap<B256, B256>>, using from_bytes_to_b256 which allows cropped values:

from_bytes_to_b256

Converts a Bytes value into a B256, accepting inputs that are less than 32 bytes long. These inputs will be left padded with zeros.