Struct reth_rpc_types::CallRequest

pub struct CallRequest {

    pub from: Option<Address>,
    pub to: Option<Address>,
    pub gas_price: Option<Uint<256, 4>>,
    pub max_fee_per_gas: Option<Uint<256, 4>>,
    pub max_priority_fee_per_gas: Option<Uint<256, 4>>,
    pub gas: Option<Uint<256, 4>>,
    pub value: Option<Uint<256, 4>>,
    pub input: CallInput,
    pub nonce: Option<Uint<64, 1>>,
    pub chain_id: Option<Uint<64, 1>>,
    pub access_list: Option<AccessList>,
    pub max_fee_per_blob_gas: Option<Uint<256, 4>>,
    pub blob_versioned_hashes: Option<Vec<FixedBytes<32>>>,
    pub transaction_type: Option<Uint<8, 1>>,

}

Call request for eth_call and adjacent methods.

Fields
from: Option<Address>

From

to: Option<Address>

To

gas_price: Option<Uint<256, 4>>

Gas Price

max_fee_per_gas: Option<Uint<256, 4>>

EIP-1559 Max base fee the caller is willing to pay

max_priority_fee_per_gas: Option<Uint<256, 4>>

EIP-1559 Priority fee the caller is paying to the block author

gas: Option<Uint<256, 4>>

Gas

value: Option<Uint<256, 4>>

Value

input: CallInput

Transaction input data

nonce: Option<Uint<64, 1>>

Nonce

chain_id: Option<Uint<64, 1>>

chain id

access_list: Option<AccessList>

AccessList

max_fee_per_blob_gas: Option<Uint<256, 4>>

Max Fee per Blob gas for EIP-4844 transactions

blob_versioned_hashes: Option<Vec<FixedBytes<32>>>

Blob Versioned Hashes for EIP-4844 transactions

transaction_type: Option<Uint<8, 1>>

EIP-2718 type

Implementations

impl CallRequest

pub fn fee_cap(&self) -> Option<Uint<256, 4>>

Returns the configured fee cap, if any.

The returns gas_price (legacy) if set or max_fee_per_gas (EIP1559)

pub fn has_empty_blob_hashes(&self) -> bool

Returns true if the request has a blobVersionedHashes field but it is empty.