Type Alias reth_rpc_types::RichBlock

pub type RichBlock = Rich<Block>;

A Block representation that allows to include additional fields

Aliased Type

struct RichBlock {
    pub inner: Block,
    pub extra_info: BTreeMap<String, Value>,
}

Fields
inner: Block

Standard value.

extra_info: BTreeMap<String, Value>

Additional fields that should be serialized into the Block object